+++ database/__init__.py
@@ -1,1 +1,5 @@
-from .db import connect_db, create_table, get_all_events, add_event, delete_event
+from .main_db import connect_db
+
+from .events_table import create_events_table, get_all_events, add_event,delete_event
+
+from .tasks_table import create_tasks_table, get_all_tasks,add_task,delete_task
